% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\docType{package}
\name{leaderstweets-package}
\alias{leaderstweets}
\alias{leaderstweets-package}
\title{leaderstweets: Access the Executive Social Media Database}
\description{
Programmatic access to the Executive Social Media Database, a collection of 498,605 tweets and their engagement metrics from 62 heads of government and state across 34 countries between 2009 and 2023. Tables are downloaded from the project's API and cached locally, so a new data release is picked up without updating this package. Downloading a full table needs a free API key, which the project issues to researchers on request. The function names and arguments mirror the companion Python package 'leaders_tweets'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/juangomezcruces/Executive-Social-Media-Database}
  \item Report bugs at \url{https://github.com/juangomezcruces/Executive-Social-Media-Database/issues}
}

}
\author{
\strong{Maintainer}: Juan Gomez Cruces \email{juansalvadorgc@gmail.com}

}
\keyword{internal}
